+++ tests/test_compile.py
@@ -65,7 +65,15 @@ def test_compile_styled(tmp_path):
         tt(df, caption="Styled")
         .fmt(j="A", digits=2)
         .style(i="header", bold=True, color="white", background="#333")
-        .style(i=0, j="A", align="c", line="tblr", padding=(0.25, 0.5))
+        .style(
+            i=0,
+            j="A",
+            align="c",
+            line="tblr",
+            line_style="dash-dotted",
+            padding=(0.25, 0.5),
+        )
+        .style(i="header", line="b", line_style="none")
         .render("typst")
     )
     _compile(typ, tmp_path)